created a header
